+++ kitty/options/definition.py
@@ -1313,10 +1313,18 @@ this option by reloading the config is not supported.
 opt('background_tint', '0.0',
     option_type='unit_float', ctype='float',
     long_text='''
-How much to tint the background image by the background color.  This option
+How much to tint the background image by the background color. This option
 makes it easier to read the text. Tinting is done using the current background
 color for each window. This option applies only if :opt:`background_opacity` is
 set and transparent windows are supported or :opt:`background_image` is set.
+'''
+    )
+
+opt('background_tint_gaps', '1.0',
+    option_type='unit_float', ctype='float',
+    long_text='''
+How much to tint the background image at the window gaps by the background
+color, after applying :opt:`background_tint`.
 '''
     )